pyproject.toml
[tool.poetry]
name = "doxxcaster"
version = "0.1.0"
description = ""
authors = ["Eshaan Bansal <eshaan7bansal@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "doxxcaster" }]

[tool.poetry.dependencies]
python = "^3.10"
farcaster = "^0.7.11"
airstack = "^0.0.7"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
